Various bug fixes, do not merge into main

+++ src/main/java/model/BookRepository.java
@@ -8,5 +8,5 @@ public interface BookRepository extends CrudRepository <Book, Integer>{
 
     List<Book> findByBookName(String bookName);
 
-    Book findById(int id);
+    Book findById(Long id);
 }
